requirements.txt
streamlit
huggingface_hub
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
PyPDF2
python-docx
